Move nodes and visitor into Arel plugin file
