+++ lib/mobility/plugins/arel.rb
@@ -1,6 +1,4 @@
 # frozen-string-literal: true
-require_relative "./arel/nodes"
-require_relative "./arel/visitor"
 
 module Mobility
   module Plugins
@@ -52,6 +50,72 @@ Arel-specific classes and modules.
           @locale = locale
           @attribute_name = attribute_name || column_name
           super(relation, column_name)
+        end
+      end
+
+      class Visitor < ::Arel::Visitors::Visitor
+        INNER_JOIN = ::Arel::Nodes::InnerJoin
+        OUTER_JOIN = ::Arel::Nodes::OuterJoin
+
+        attr_reader :backend_class, :locale
+
+        def initialize(backend_class, locale)
+          super()
+          @backend_class, @locale = backend_class, locale
+        end
+
+        private
+
+        def visit(*args)
+          super
+        rescue TypeError
+          visit_default(*args)
+        end
+
+        def visit_collection(_objects)
+          raise NotImplementedError
+        end
+        alias :visit_Array :visit_collection
+
+        def visit_Arel_Nodes_Unary(object)
+          visit(object.expr)
+        end
+
+        def visit_Arel_Nodes_Binary(object)
+          visit_collection([object.left, object.right])
+        end
+
+        def visit_Arel_Nodes_Function(object)
+          visit_collection(object.expressions)
+        end
+
+        def visit_Arel_Nodes_Case(object)
+          visit_collection([object.case, object.conditions, object.default])
+        end
+
+        def visit_Arel_Nodes_And(object)
+          visit_Array(object.children)
+        end
+
+        def visit_Arel_Nodes_Node(object)
+          visit_default(object)
+        end
+
+        def visit_Arel_Attributes_Attribute(object)
+          visit_default(object)
+        end
+
+        def visit_default(_object)
+          nil
+        end
+      end
+
+      module Nodes
+        class Binary < ::Arel::Nodes::Binary; end
+        class Grouping < ::Arel::Nodes::Grouping; end
+
+        ::Arel::Visitors::ToSql.class_eval do
+          alias :visit_Mobility_Plugins_Arel_Nodes_Grouping :visit_Arel_Nodes_Grouping
         end
       end
     end
